scripts/blog-bot/pyproject.toml
[project]
name = "gridmind-blog-bot"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.40.0",
    "httpx>=0.27.0",
    "PyGithub>=2.3.0",
    "python-frontmatter>=1.1.0",
    "schedule>=1.2.0",
    "structlog>=24.1.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100
